+++ mint/internal/lightning/util.go
@@ -1,0 +1,1 @@
+package lightning
